Separate income variables into separate files
