+++ openfisca_us/variables/income/household/household_net_income.py
@@ -1,5 +1,4 @@
-from openfisca_tools.model_api import *
-from openfisca_us.entities import *
+from openfisca_us.model_api import *
 
 
 class household_net_income(Variable):
@@ -7,6 +6,7 @@ class household_net_income(Variable):
     entity = Household
     label = u"Household net income"
     definition_period = YEAR
+    unit = "currency-USD"
 
     def formula(household, period, parameters):
         return household.sum(household.members("net_income", period))
