+++ openfisca_us/variables/income/person/market_income.py
@@ -1,0 +1,10 @@
+from openfisca_us.model_api import *
+
+
+class market_income(Variable):
+    value_type = float
+    entity = Person
+    label = "Market income"
+    unit = "currency-USD"
+    documentation = "Income from all non-government sources"
+    definition_period = YEAR
