+++ openfisca_us/variables/income/person/net_income.py
@@ -8,12 +8,3 @@ class net_income(Variable):
     unit = "currency-USD"
     documentation = "Personal disposable income after taxes and transfers"
     definition_period = YEAR
-
-
-class market_income(Variable):
-    value_type = float
-    entity = Person
-    label = "Market income"
-    unit = "currency-USD"
-    documentation = "Income from all non-government sources"
-    definition_period = YEAR
