+++ openfisca_us/variables/income/spm_unit/spm_unit_total_income.py
@@ -1,0 +1,9 @@
+from openfisca_us.model_api import *
+
+
+class spm_unit_total_income(Variable):
+    value_type = float
+    entity = SPMUnit
+    label = u"SPM unit total income"
+    definition_period = YEAR
+    unit = "currency-USD"
