+++ openfisca_us/variables/income/spm_unit/spm_unit_wic.py
@@ -1,0 +1,9 @@
+from openfisca_us.model_api import *
+
+
+class spm_unit_wic(Variable):
+    value_type = float
+    entity = SPMUnit
+    label = u"SPM unit WIC subsidy"
+    definition_period = YEAR
+    unit = "currency-USD"
